Improve symlink creation checking
Symlinks will only be created if they do not already exist. If they do
exist and already point to the correct target, nothing is done. If they
point to the wrong target, or already exist and are just regular files
then the user will be asked whether to replace the file/link.